+++ common/ShotMessage.cpp
@@ -12,7 +12,7 @@ const StaticString shotKey("shot");
 
 } // end anonymous namespace
 
-ShotMessage::ShotMessage(int id, const Shot& s) :
+ShotMessage::ShotMessage(uint16_t id, const Shot& s) :
 	Message(id),
 	shot(s)
 {
