+++ common/TargetControlMessage.cpp
@@ -33,16 +33,16 @@ TargetCommand TargetCommand::fromJSON(const Json::Value& object)
 	ENFORCE(IOException, idValue.isInt(), "The target ID is not an integer.");
 	ENFORCE(IOException, onValue.isBool(), "The target ID is not a bool.");
 
-	return TargetCommand(idValue.asInt(), onValue.asBool());
+	return TargetCommand((int8_t)idValue.asInt(), onValue.asBool());
 }
 
-TargetControlMessage::TargetControlMessage(int id, CommandList&& comms) :
+TargetControlMessage::TargetControlMessage(uint16_t id, CommandList&& comms) :
 	Message(id),
 	commands(move(comms))
 {
 }
 
-TargetControlMessage::TargetControlMessage(int id, const TargetCommand& comm) :
+TargetControlMessage::TargetControlMessage(uint16_t id, const TargetCommand& comm) :
 	Message(id),
 	commands({comm})
 {
