+++ common/TestMessage.hpp
@@ -7,7 +7,7 @@ class TestMessage : public Message {
 
 public:
 
-	TestMessage(int id, const Json::Value& object = Json::Value());
+	TestMessage(uint16_t id, const Json::Value& object = Json::Value());
 
 	static std::unique_ptr<TestMessage> fromJSON(const Json::Value& object);
